+++ py_example/data_processing/PathMinerLoader.py
@@ -1,9 +1,8 @@
 import pandas as pd
-import numpy as np
 from os import path
 import parse
 
-from UtilityEntities import PathContext, Path, NodeType
+from data_processing.UtilityEntities import PathContext, Path, NodeType
 
 
 # Loads all .csv files generated by PathMiner in pd.Series or pd.Dataframe
